III-5746 - Rename cultuurkuurstep and imports
